requirements.txt
Flask==2.0.1
requests==2.25.1
beautifulsoup4==4.9.3
python-dateutil==2.8.1
lxml==4.6.3
